# Remove array/object allocs during Alt+Numpad composition, improve performance