--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/TextCompositionManager.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/TextCompositionManager.cs
@@ -1,11 +1,9 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Runtime.InteropServices;
 using System.Windows.Threading;
-using System.Diagnostics;
+using System.ComponentModel;
 using Microsoft.Win32;
-using System.Text;
 using MS.Win32;
 
 namespace System.Windows.Input
@@ -490,7 +488,7 @@ namespace System.Windows.Input
                 KeyEventArgs keyArgs = (KeyEventArgs)e.StagingItem.Input;
                 if (!keyArgs.Handled)
                 {
-                    if (keyArgs.RealKey == Key.LeftAlt || keyArgs.RealKey == Key.RightAlt)
+                    if (keyArgs.RealKey is Key.LeftAlt or Key.RightAlt)
                     {
                         // Make sure both Alt keys are up.
                         ModifierKeys modifiers = keyArgs.KeyboardDevice.Modifiers;
@@ -505,16 +503,15 @@ namespace System.Windows.Input
                                 if (_altNumpadEntry != 0)
                                 {
                                     _altNumpadcomposition.ClearTexts();
-                                    if (_altNumpadConversionMode == AltNumpadConversionMode.OEMCodePage)
+                                    if (_altNumpadConversionMode is AltNumpadConversionMode.OEMCodePage)
                                     {
                                         _altNumpadcomposition.SetText(GetCurrentOEMCPEncoding(_altNumpadEntry));
                                     }
-                                    else if ((_altNumpadConversionMode == AltNumpadConversionMode.DefaultCodePage) ||
-                                             (_altNumpadConversionMode == AltNumpadConversionMode.HexDefaultCodePage))
+                                    else if (_altNumpadConversionMode is AltNumpadConversionMode.DefaultCodePage or AltNumpadConversionMode.HexDefaultCodePage)
                                     {
                                         _altNumpadcomposition.SetText(CharacterEncoding(InputLanguageManager.Current.CurrentInputLanguage.TextInfo.ANSICodePage, _altNumpadEntry));
                                     }
-                                    else if (_altNumpadConversionMode == AltNumpadConversionMode.HexUnicode)
+                                    else if (_altNumpadConversionMode is AltNumpadConversionMode.HexUnicode)
                                     {
                                         _altNumpadcomposition.SetText(((char)_altNumpadEntry).ToString());
                                     }