## Description

Removes unnecessary `byte[]` , `char[]` and `StringBuilder` alocations during post-input processing when the char is inputted via Alt+Numpad combination. This is not somewhat impactful allocation but it will allow for more optimizations in the future.

Time spent is cut by 2/5 which is more impactful in this case.

### GetCurrentOEMCPEncoding method chain call

| Method   |Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |-----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | 48.77 ns |   0.674 ns |    0.630 ns | 0.0086 |         814 B |         144 B |
| PR_EDIT  |  31.83 ns |   0.292 ns |    0.273 ns | 0.0019 |         418 B |          32 B |

## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build, text input.

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9954)